numbers: Use currency decimal digits by default
When formatting a price, the number of decimals to use should be defined
by the currency.  For example, JPY do not use decimal numbers at all.

However, we still allow the user to override the currency decimal digits
setting with the one from the number format.

Fixes https://github.com/mitsuhiko/babel/issues/139