Resolve locale data aliases eagerly
Get rid of the LocaleDataDict because it slows down key lookups compared to a
native dict.  To do that, we need to resolve the aliases when loading the locale
data from disk.

Since this data is later cached, we incur in a slight overhead for the initial
load.  However, this is well compensated during the rest of the process
execution.