Fixed issue #109: `ImportWarning` warned when `import babel`
`localedata.py` uses the same name with folder `localedata/`, which
Python will try to load `localdata/__init__.py` and warn `ImportWarning`
if the file not found.

If try to filter the `ImportWarning` as an error like this:

```python
import warnings
warnings.filterwarnings('error', category=ImportWarning)
import babel
```

An `ImportWarning` exception will be raised.