Add rultor configuration
This allows merging via github comments. The comment `@rultor merge`
will execute the script (which currently doesn't do anything) and if it
succeeds will perform the merge.

I plan to use rultor later for:
 * Automatic deployment to PyPI (development releases directly from
   master, we do this already in coala, see
   https://github.com/coala-analyzer/coala/blob/master/.rultor.yml)
 * Automatic releasing with deployment to PyPI.
 * Veryfy that all CI services pass before merging (see
   https://github.com/yegor256/rultor/issues/869)

With this any manual pushes to master are disallowed, all pushes to
master have to be validated by continuous integration and reviewed by a
non-committer.