Make catalog header updating an option
The change in e0e7ef168856bb had an unexpected and likely undesired effect
when updating catalogs with, e.g. translator names in the header comment.

It's better to make the updating an option and revert back to the pre-2.2
behavior by default.

Fixes https://github.com/python-babel/babel/issues/318