Increase coverage of JSX extraction tests
b5e0ad8 fixed a few bugs with jsx parsing but only added
tests for the lexer. This commit adds a test for the extraction
which did not work before:

    <tag>{ _"Text" }</tag>

This test case is not present yet in the suite, only tags with
attributes (<tag attr="value">) are, but they get extracted correctly
already.