Add format argument to ``parse_date()``
``parse_date()`` does not allow the user to specify the date's format
and the 'medium' format is used by default in the call to
``get_date_format()`` on line 1144. This results in a failure to parse
the date in short format for the locale 'sv_SE'. This commit adds the
format argument to avoid this kind of failures. The default value is
set to 'medium' to preserve the old behavior.

Closes https://github.com/python-babel/babel/issues/657