Better fix overflow of ISO week numbers
PR #621 had added an adjustment for overflowing week
numbers (e.g. days of calendar year A living in weak-year A+1),
however this adjustment would mis-trigger on *underflow*: in that case
the year of `day_of_period` and `self.value` differ (`self.value` is
the next year), meaning the number of weeks in the year differs, so
when underflowing a week from a 52-weeks year to a 53-weeks year, the
weeks count would then be overflowed back, and associated with the
wrong week-year.

Example: in 2010-01-04 is on a Monday, so 2010-01-03 is part of the
last week of 2009, which is 2009 W53.

But because of the way `get_week_number` works we first compute the
week number of "3, sunday", find that it's 0, compute the week number
of December 31st, 2009, get 53, then check the number of weeks
against *2010* because `isocalendar` is called with 2010-01-03
adjusted to *2010*-12-28, thus we end up with a week number of 1,
associated with a correct week-year of 2009.

Hence being told that 2010-01-03 is in 2009-W01.

Fix by moving the modulo operation to `format_week` and only to the
case where we may have overflowed the year.

Notes:

- the check is incredibly gross
- the issue of overflowing the year occurs in every week numbering
  system, however it means we need to compute the number of weeks in
  non-ISO locales
- also it's *under*flowing which is only an issue in ISO and
  technically it's not even ISO itself, but a `min_week_days` which is
  not 1, otherwise by definition the week-year of a date can only be
  equal to or higher than the calendar year
- similar issues almost certainly occurs with months as well, but I'm
  not convinced it matters to anyone given week-months are not a thing
  so non-split month-weeks are unusable