Keep @modifiers when parsing locales
Locale modifiers ("@variants") are described in the GNU gettext
documentation like this:

> The ‘@variant’ can denote any kind of characteristics that is not
> already implied by the language ll and the country CC. […] It can also
> denote a dialect of the language, …

Wherein Babel previously would discard these, this patch stores the
modifier information in the `Locale` objects, handling string
representation accordingly.

Not implemented is the lookup of a meaningful description of modifiers,
but instead — for now — an identity mapping is provided.

Resolves: #946
Signed-off-by: martin f. krafft <madduck@madduck.net>