Freeze format_time() tests to a specific date to fix test failures
Freeze the date when performing the tests for format_time() with
a timezone specified.  Since the time object does not specify a date,
the formatter uses the format string specific to the current date.
As a result, if the current DST state is different than when the test
was last updated, it failed.

This fix covers only regular tests.  I have no idea how to do the same
for doctests.

Issue #988